[CoreBundle][UserBundle] Move DefaultUsernameORMListener to core [CoreBundle] Fix broken 'use'
